Moved shooting logic into entityBiped class
