+++ levels/test.txt
@@ -4,7 +4,7 @@ background "jungle2" 0 0 1067 527 0
 let surf1 : surface (-200) (-200) 200 (-200)
 surface 200 (-200) 200 (-150)
 surface (-200) (-300) (-200) (-200)
-surface (-190) (-200) (-190) (-300)
+surface (-190) (-203) (-190) (-300)
 surface (-187) (-300) (-203) (-300)
 surface 200 (-150) 300 (-150)
