+++ player.py
@@ -7,8 +7,6 @@ from entity.biped import EntityBiped
 class Player(EntityBiped):
     def __init__(self, level, pos):
         super().__init__(level, pos)
-
-        self.level = level
         
         self.spawnpoint = Vec(*pos)
         self.walljump = False
@@ -61,15 +59,10 @@ class Player(EntityBiped):
         elif self.walljump:
             self.wallJumpTime += 1
 
-        if self.touching_platform:
-            self.vel.x *= 0.6
-        else:
-            self.vel.x *= 0.925
-
         if abs(self.vel.x) < 0.01:
             self.vel.x = 0
 
-        if False: ## FLIGHT CHECK HERE TODO ??
+        if False: ## FLIGHT CHECK HERE TODO ?? - this would be the cheat-flying mode for debugging, maybe remove?
             if keys[controlsMap["up"]]:
                 self.yVel = 5
             elif keys[controlsMap["sneak"]]:
@@ -91,31 +84,13 @@ class Player(EntityBiped):
             self.level.set_player(self)
             self.level.start()
         
-        if controls["shoot"] and self.gun_cooldown == 0:
-            if self.power >= 40 and not self.reload: ## TODO - these magic numbers are gunPower
-                controller.sound_ctrl.play_sound(controller.sounds["player_shoot"])
-                bulletspeed = 20
-                self.gun_cooldown = 20
-                self.level.projectiles.append(Bullet(*self.view.held_pos, self.view.aim, bulletspeed, self))
-                self.power -= 40 ## TODO - magic number
-                if self.power < 40:
-                    self.reload = True
-            else:
-                controller.sound_ctrl.play_sound(controller.sounds["click"])
-                self.gun_cooldown = 30
+        if controls["shoot"]:
+            self.shoot()
 
         if self.gun_cooldown > 0:
             self.gun_cooldown -= 1
         if controls["reload"] and self.gun_cooldown == 0:
                 self.reload = True
-
-        if self.reload:
-            self.power += self.reload_speed
-            if self.power > self.max_power:
-                self.power = self.max_power
-            if self.power == self.max_power:
-                self.reload = False
-                self.gun_cooldown = 0
 
         if self.pos.y < -2000:
             self.hp = 0
@@ -128,81 +103,3 @@ class Player(EntityBiped):
 
     def render(self, camera_pos):
         super().render(camera_pos)
-
-class Bullet:
-    def __init__(self, x, y, d, speed, owner):
-        self.x = x
-        self.y = y
-        self.d = d
-        self.owner = owner
-        self.xVel = (Cos(d)*(speed))#+owner.xVel
-        self.yVel = (Sin(d)*(speed))#+owner.yVel
-        self.speed = speed
-        self.age = 0
-
-    def update(self):
-        self.x += self.xVel
-        self.y += self.yVel
-        self.age += 1
-        #self.yVel -= 1
-
-        return 0 < self.age < 100
-
-    def check_forward(self, pr, platform):
-        if (
-            self.x + (self.xVel*pr) < platform.x + (platform.w/2) and
-            self.x + (self.xVel*pr) > platform.x - (platform.w/2) and
-            self.y + (self.yVel*pr) > platform.y - (platform.h/2) and
-            self.y + (self.yVel*pr) < platform.y + (platform.h/2)
-        ):
-            return True
-        else:
-            return False
-
-    def get_touching(self, level):
-        self.touchingPlatform = False
-        self.rightTouching = 0
-        self.leftTouching = 0
-        self.upTouching = 0
-        self.downTouching = 0
-
-        for platform in level.platforms:
-            if platform.d == 0:
-                if (
-                    self.x < platform.x + (platform.w/2) and
-                    self.x > platform.x - (platform.w/2) and
-                    self.y > platform.y - (platform.h/2) and
-                    self.y < platform.y + (platform.h/2)
-                ):
-                    return True
-                elif distance(self.x, self.y, platform.x, platform.y)<=max(platform.w,platform.h):
-                    pr = -1
-
-                    while abs(pr*self.speed)>2:
-                        pr *= 0.9
-
-                        if self.check_forward(pr, platform):
-                            return True
-            else:
-                x3 = platform.x - ((platform.w/2) * Cos(-platform.d)) + ((platform.h/2) * Sin(-platform.d))
-                y3 = platform.y + ((platform.h/2) * Cos(-platform.d)) + ((platform.w/2) * Sin(-platform.d))
-                x4 = platform.x + ((platform.w/2) * Cos(-platform.d)) + ((platform.h/2) * Sin(-platform.d))
-                y4 = platform.y + ((platform.h/2) * Cos(-platform.d)) - ((platform.w/2) * Sin(-platform.d))
-                col = line_collision((self.x, self.y, self.x-self.xVel, self.y-self.yVel),
-                                     (x3, y3, x4, y4))
-                if col[0]:
-                    return True
-
-        for entity in level.entities:
-            if not entity == self.owner:
-                hit, xD, yD = entity.check_inside(self.x, self.y)
-                if hit:
-                    #print(hit, xD, yD)
-                    entity.damage(1, 2, (self.xVel*0.2, self.yVel*0.2 + 1))
-                    return True
-        if not level.player == self.owner:
-            hit, xD, yD = level.player.check_inside(self.x, self.y)
-            if hit:
-                level.player.damage(1, 2, (self.xVel*1, self.yVel*1 + 1))
-                return True
-        return False
